Add Sphinx for generating documentation
Why these changes are being introduced:
Sphinx is the standard Python tool for generating detailed
documentation from docstrings. We will use it to generate and update the
official docs as we develop and maintain the library.

How this addresses that need:
* Adds Sphinx, theme, and extras to dev dependencies
* Creates a docs folder with initial Sphinx config and files
* Updates README format for easy inclusion in Sphinx-generated docs
* Creates a make command to build docs with Sphinx
* Updates PR template to include Sphinx in developer checklist

Side effects of this change:
None, but a new ticket has been created to handle publishing the docs
(ETD-420, linked below).

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-392
* https://mitlibraries.atlassian.net/browse/ETD-420