Add get_object_by_handle method to DSpaceClient
Why these changes are being introduced:
The DSpace API requires object UUIDs to post to Communities,
Collections, and Items, however other applications generally store and
reference those objects by their handles. The API provides a generic
endpoint to retrieve an object by its handle and this commit implements
that in the client. This will allow applications using the library to
post to objects by UUID even if they only have the object handles
(which will often be the case). It's also a useful precursor to
implementing the Item class and post method.

How this addresses that need:
* Adds a get_object_by_handle method to the DSpaceClient class, along
  with corresponding unit tests and VCR cassettes.
* Adds an authenticated test_client fixture to simplify tests that
  require an authentication cookie.
* Updates VCR response scrubbing to remove additional response body
  content.

Side effects of this change:
* Adds optional json param to client post method to take advantage of
  requests' built-in JSON encoder
* Updates README to reflect changes to recreation of VCR cassettes with
  new test_client fixture
* Updates documentation in the dspace.client module for consistency
* Updates sphinx config