Add Item module and tests
Why these changes are being introduced:
We need models for DSpace Item and MetadataEntry objects, along with
associated API calls for the /items endpoint. The initial use case for
this library requires the ability to create new items in DSpace, so
that's the first method added.

How this addresses that need:
* Adds a dspace.item module with classes for Item and MetadataEntry
  objects
* Adds a post method to the Item class
* Adds ability to create MetadataEntry objects from dicts and serialize
  them to JSON for IO operations
* Adds tests for new classes and methods

Side effects of this change:
* Updates Sphinx source docs to include new module

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-392