Linting and documentation updates
Why these changes are being introduced:
Mypy and Pydocstyle are static analysis tools checking type hints and
docstrings, respectively. Since we are using both extensively in this
library, we should run automated checks on them to ensure we are
following the standards.

How this addresses that need:
* Adds mypy and pydocstyle to dev dependencies, Makefile, and CI
* Adds mypy config to pyproject.toml
* Updates existing modules to fix errors found by these tools

Side effects of this change:
* Removes CI on pull request since it's redundant when we check on each
  push
* Updates an incorrect VCR cassette path in test_client.py and deletes
  the extra cassette
* Updates dependencies

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-405