Configure Sphinx docs for publishing on Github
Why these changes are being introduced:
We want to publish the docs in a human-readable format. Using Github
Pages is an easy way to do this directly from the main branch of the
repo, without having to configure an external service like readthedocs.
However, Github pages expects a Jekyll site by default so the Sphinx
built docs need to be reconfigured a bit to allow GH pages to publish
them.

How this addresses that need:
* Updates Sphinx Makefile to create build files in the docs/html folder
* Creates an index.html file in the docs folder that redirects to the
  Sphinx-built index.html file in docs/html/html
* Adds a .nojekyll file to the docs folder so GH pages knows it isn't a
  jekyll site

Side effects of this change:
If the Github repo settings are configured correctly, all merges into
the main branch will publish updates docs to the repo's Github Pages
site.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-420